Increase level when experience >= nextLevelExperience

+++ enc_temp_folder/b2388e1501bfe749d5e8b7e8d59c6ca/BaseCharacter.cpp
@@ -103,9 +103,9 @@ float ABaseCharacter::GetAttackPower() const
 	return Attributes->GetAttackPower();
 }
 
-int ABaseCharacter::GetLevel() const
+float ABaseCharacter::GetLevel() const
 {
-	return static_cast<int>(Attributes->GetLevel());
+	return Attributes->GetLevel();
 }
 
 float ABaseCharacter::GetExperience() const
